backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
passlib==1.7.4
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
